Moved connections to internal package, config to internal pakage, and extension to internal package
