+++ src/main/java/com/avioconsulting/mule/vault/provider/api/connection/parameters/AWSIdentityProperties.java
@@ -1,6 +1,6 @@
 package com.avioconsulting.mule.vault.provider.api.connection.parameters;
 
-import com.avioconsulting.mule.vault.provider.api.VaultPropertiesProviderExtension;
+import com.avioconsulting.mule.vault.provider.internal.extension.VaultPropertiesProviderExtension;
 import org.mule.runtime.api.meta.ExpressionSupport;
 import org.mule.runtime.config.api.dsl.model.ConfigurationParameters;
 import org.mule.runtime.extension.api.annotation.Expression;
