+++ src/main/java/com/avioconsulting/mule/vault/provider/internal/config/VaultPropertiesProviderConfiguration.java
@@ -1,6 +1,6 @@
-package com.avioconsulting.mule.vault.provider.api;
+package com.avioconsulting.mule.vault.provider.internal.config;
 
-import com.avioconsulting.mule.vault.provider.api.connection.provider.*;
+import com.avioconsulting.mule.vault.provider.internal.connection.provider.*;
 import org.mule.runtime.extension.api.annotation.Configuration;
 import org.mule.runtime.extension.api.annotation.connectivity.ConnectionProviders;
