+++ src/main/java/com/avioconsulting/mule/vault/provider/internal/connection/impl/AbstractConnection.java
@@ -1,6 +1,6 @@
-package com.avioconsulting.mule.vault.provider.api.connection.impl;
+package com.avioconsulting.mule.vault.provider.internal.connection.impl;
 
-import com.avioconsulting.mule.vault.provider.api.connection.VaultConnection;
+import com.avioconsulting.mule.vault.provider.internal.connection.VaultConnection;
 import com.avioconsulting.mule.vault.provider.api.connection.parameters.SSLProperties;
 import com.bettercloud.vault.SslConfig;
 import com.bettercloud.vault.Vault;
