+++ src/main/java/com/avioconsulting/mule/vault/provider/internal/connection/provider/AbstractAWSConnectionProvider.java
@@ -1,4 +1,4 @@
-package com.avioconsulting.mule.vault.provider.api.connection.provider;
+package com.avioconsulting.mule.vault.provider.internal.connection.provider;
 
 import org.mule.runtime.config.api.dsl.model.ConfigurationParameters;
 import org.mule.runtime.extension.api.annotation.param.Optional;
