+++ src/test/java/com/avioconsulting/mule/vault/provider/internal/connection/parameters/EC2PropertiesTestCase.java
@@ -1,5 +1,7 @@
-package com.avioconsulting.mule.vault.provider.api.connection.parameters;
+package com.avioconsulting.mule.vault.provider.internal.connection.parameters;
 
+import com.avioconsulting.mule.vault.provider.api.connection.parameters.AWSIdentityProperties;
+import com.avioconsulting.mule.vault.provider.api.connection.parameters.EC2ConnectionProperties;
 import junit.framework.TestCase;
 import org.junit.Test;
