+++ src/test/java/com/avioconsulting/mule/vault/provider/internal/connection/parameters/TLSPropertiesTestCase.java
@@ -1,5 +1,8 @@
-package com.avioconsulting.mule.vault.provider.api.connection.parameters;
+package com.avioconsulting.mule.vault.provider.internal.connection.parameters;
 
+import com.avioconsulting.mule.vault.provider.api.connection.parameters.JKSProperties;
+import com.avioconsulting.mule.vault.provider.api.connection.parameters.PEMProperties;
+import com.avioconsulting.mule.vault.provider.api.connection.parameters.TLSAuthProperties;
 import junit.framework.TestCase;
 import org.junit.Test;
